refactor for state co location
